adapts CheckGreatest to use PT interface; reorganizes lemmas by moving them into the utils package
