+++ pkg/utils/bytes.gobra
@@ -19,7 +19,7 @@ ghost
 requires acc(BytesMem(s), _)
 decreases
 opaque
-pure func getContent(s []byte) (res seq[byte]) {
+pure func getContent(s []byte) seq[byte] {
   return unfolding acc(BytesMem(s), _) in getByteContent(s, 0)
 }
 
@@ -27,6 +27,6 @@ ghost
 requires acc(s)
 requires 0 <= idx && idx <= len(s)
 decreases len(s) - idx
-pure func getByteContent(s []byte, idx int) (res seq[byte]) {
+pure func getByteContent(s []byte, idx int) seq[byte] {
   return idx == len(s) ? seq[byte]{} : seq[byte]{s[idx]} ++ getByteContent(s, idx + 1)
 }
