+++ pkg/utils/hash.gobra
@@ -1,0 +1,18 @@
+package utils
+
+import "crypto/sha256"
+
+ghost
+requires acc(s, _)
+decreases
+opaque
+pure func GetHashContent(s *[sha256.Size]byte) seq[byte] {
+    return getHashContentRec(*s, 0)
+}
+
+ghost
+requires 0 <= idx && idx <= sha256.Size
+decreases sha256.Size - idx
+pure func getHashContentRec(s [sha256.Size]byte, idx int) seq[byte] {
+  return idx == sha256.Size ? seq[byte]{} : seq[byte]{s[idx]} ++ getHashContentRec(s, idx + 1)  
+}
